requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
dj-database-url>=2.1.0
drf-yasg